// packages/mpl-core-kit-lib/src/generated/types/baseLifecycleHook.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  type Address,
  type Codec,
  combineCodec,
  type Decoder,
  type Encoder,
  getAddressDecoder,
  getAddressEncoder,
  getArrayDecoder,
  getArrayEncoder,
  getOptionDecoder,
  getOptionEncoder,
  getStructDecoder,
  getStructEncoder,
  type Option,
  type OptionOrNullable,
} from '@solana/kit'

import {
  type BaseExtraAccount,
  type BaseExtraAccountArgs,
  type BasePluginAuthority,
  type BasePluginAuthorityArgs,
  type ExternalPluginAdapterSchema,
  type ExternalPluginAdapterSchemaArgs,
  getBaseExtraAccountDecoder,
  getBaseExtraAccountEncoder,
  getBasePluginAuthorityDecoder,
  getBasePluginAuthorityEncoder,
  getExternalPluginAdapterSchemaDecoder,
  getExternalPluginAdapterSchemaEncoder,
} from './index.ts'

export type BaseLifecycleHook = {
  dataAuthority: Option<BasePluginAuthority>
  extraAccounts: Option<Array<BaseExtraAccount>>
  hookedProgram: Address
  schema: ExternalPluginAdapterSchema
}

export type BaseLifecycleHookArgs = {
  dataAuthority: OptionOrNullable<BasePluginAuthorityArgs>
  extraAccounts: OptionOrNullable<Array<BaseExtraAccountArgs>>
  hookedProgram: Address
  schema: ExternalPluginAdapterSchemaArgs
}

export function getBaseLifecycleHookCodec(): Codec<BaseLifecycleHookArgs, BaseLifecycleHook> {
  return combineCodec(getBaseLifecycleHookEncoder(), getBaseLifecycleHookDecoder())
}

export function getBaseLifecycleHookDecoder(): Decoder<BaseLifecycleHook> {
  return getStructDecoder([
    ['hookedProgram', getAddressDecoder()],
    ['extraAccounts', getOptionDecoder(getArrayDecoder(getBaseExtraAccountDecoder()))],
    ['dataAuthority', getOptionDecoder(getBasePluginAuthorityDecoder())],
    ['schema', getExternalPluginAdapterSchemaDecoder()],
  ])
}

export function getBaseLifecycleHookEncoder(): Encoder<BaseLifecycleHookArgs> {
  return getStructEncoder([
    ['hookedProgram', getAddressEncoder()],
    ['extraAccounts', getOptionEncoder(getArrayEncoder(getBaseExtraAccountEncoder()))],
    ['dataAuthority', getOptionEncoder(getBasePluginAuthorityEncoder())],
    ['schema', getExternalPluginAdapterSchemaEncoder()],
  ])
}